Add patch to allow searching fields in a struct by tags.
I don't think this will ever make it to upstream because it's very hackish.
It allows to keep backwards compatible structs for template execution.

Example:

You have this struct:

```go
type config struct {
	Dns []string
}
```

Templates can search by the `Dns` field by using `{{.Dns}}`.

One day, you decide to apply go-lint that gives you a warning because
`Dns` should be `DNS` and you decide to change it:

```go
type config struct {
	DNS []string
}
```

All the previous templates will break with this change. This might be
okay if you control those templates, but it's not that good if you allow
consumers of your software to provide their own templates, like docker
does with their inspect command in their CLI:

```
$ docker inspect --format "{{.Dns}}" container_id
```

This patch allows to keep the old field name in a struct tag. The
template will search for the name in the field tags before failing the
execution because a field does not exist:

```go
type config struct {
	DNS []string `template:"Dns"`
}
```

The performance penalty might be very bad if you have a lot of fields
declared with this tag. You should decide when it makes sense to break
backwards compatibility with consumers before abusing this tag.

Signed-off-by: David Calavera <david.calavera@gmail.com>